+++ AutomatedLab/AutomatedLab.psm1
@@ -3829,6 +3829,31 @@ function Clear-LabCache
 }
 #endregion Clear-LabCache
 
+#region Get-LabCache
+function Get-LabCache
+{
+    [CmdletBinding()]
+    param
+    ( )
+
+    $regKey = [Microsoft.Win32.RegistryKey]::OpenBaseKey('CurrentUser', 'Default')
+    try
+    {
+        $key = $regKey.OpenSubKey('Software\AutomatedLab\Cache')
+        foreach ($value in $key.GetValueNames())
+        {
+            $content = [xml]$key.GetValue($value)
+            $timestamp = $content.SelectSingleNode('//Timestamp')
+            [pscustomobject]@{
+                Store = $value
+                Timestamp = $timestamp.datetime -as [datetime]
+            }
+        }
+    }
+    catch { Write-Verbose -Message "Cache not yet created" }
+}
+#endregion
+
 #region function Add-LabVMUserRight
 function Add-LabVMUserRight
 {